+++ eminus/scf.py
@@ -76,7 +76,7 @@ class SCF:
         # Set min here, better not use mutable data types in signatures
         if self.min is None:
             # For systems with bad convergence throw in some sd steps
-            self.min = {'sd': 25, 'pccg': 250}
+            self.min = {'auto': 250}
 
         # Initialize logger
         self.log = create_logger(self)