+++ tests/test_localizer.py
@@ -14,7 +14,7 @@ atoms_unpol = Atoms('CH4', ((0, 0, 0),
                             (1.186, -1.186, -1.186),
                             (-1.186, 1.186, -1.186),
                             (-1.186, -1.186, 1.186)), center=True, ecut=5, Nspin=1)
-scf_unpol = SCF(atoms_unpol, min={'pccg': 15})
+scf_unpol = SCF(atoms_unpol)
 scf_unpol.run()
 
 atoms_pol = Atoms('CH4', ((0, 0, 0),
@@ -22,7 +22,7 @@ atoms_pol = Atoms('CH4', ((0, 0, 0),
                           (1.186, -1.186, -1.186),
                           (-1.186, 1.186, -1.186),
                           (-1.186, -1.186, 1.186)), center=True, ecut=5, Nspin=2)
-scf_pol = SCF(atoms_pol, min={'sd': 3, 'pccg': 15})
+scf_pol = SCF(atoms_pol)
 scf_pol.run()
 
 # FODs that will be used for both spin channels