+++ tests/test_minimizer.py
@@ -20,7 +20,7 @@ scf_pol = SCF(atoms_pol, etol=tolerance)
 def test_minimizer_unpol(minimizer):
     '''Check the spin-unpaired minimizer functions.'''
     scf_unpol.min = {minimizer: 100}
-    if 'cg' in minimizer:
+    if minimizer in ('cg', 'pccg', 'auto'):
         for i in range(1, 4):
             scf_unpol.cgform = i
             E = scf_unpol.run()
@@ -34,7 +34,7 @@ def test_minimizer_unpol(minimizer):
 def test_minimizer_pol(minimizer):
     '''Check the spin-paired minimizer functions.'''
     scf_pol.min = {minimizer: 100}
-    if 'cg' in minimizer:
+    if minimizer in ('cg', 'pccg', 'auto'):
         for i in range(1, 4):
             scf_pol.cgform = i
             E = scf_pol.run()