+++ tests/test_tools.py
@@ -12,16 +12,17 @@ from eminus.tools import (center_of_mass, check_norm, check_ortho, check_orthono
                           get_tautf, get_tauw, gridspacing2cutoff, inertia_tensor, orbital_center)
 
 
+min = {'sd': 25, 'pccg': 25}
 atoms = Atoms('He2', ((0, 0, 0), (10, 0, 0)), s=15, Nspin=1, center=True)
-scf = SCF(atoms)
+scf = SCF(atoms, min=min)
 scf.run()
 psi = atoms.I(get_psi(scf, scf.W))
 
 atoms_unpol = Atoms('He', (0, 0, 0), ecut=1, Nspin=1)
 atoms_pol = Atoms('He', (0, 0, 0), ecut=1, Nspin=2)
-scf_unpol = SCF(atoms_unpol)
+scf_unpol = SCF(atoms_unpol, min=min)
 scf_unpol.run()
-scf_pol = SCF(atoms_pol)
+scf_pol = SCF(atoms_pol, min=min)
 scf_pol.run()
 
 
@@ -45,7 +46,7 @@ def test_center_of_mass(coords, masses, ref):
 def test_pycom(Nspin):
     '''Test PyCOM routine.'''
     atoms = Atoms('He2', ((0, 0, 0), (10, 0, 0)), s=10, Nspin=Nspin, center=True).build()
-    scf = SCF(atoms)
+    scf = SCF(atoms, min=min)
     scf.run()
     psi = atoms.I(get_psi(scf, scf.W))
     for spin in range(Nspin):